dc-due-diligence/pyproject.toml
[project]
name = "dc-due-diligence"
version = "0.1.0"
description = "Data center due diligence workflow automation -- document processing and analysis"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.0",
    "openpyxl>=3.1.0",
    "pyxlsb>=1.0.10",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "anthropic>=0.40.0",
    "Pillow>=10.0.0",
    "tavily-python>=0.7.0",
    "exa-py>=2.0.0",
    "firecrawl-py>=4.0.0",
    "apify-client>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["converters*", "research*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
